common: beerocks: change the datatype of signal_strength
change datatype of signal strength from uint8_t to int8_t so that it can
store the negative values as RSSI will be in negative.

PPM-3389

(cherry picked from commit fe1c8009699197d8ac32b6d498fb102eb9c1caee)

+++ common/beerocks/bwl/include/bwl/base_wlan_hal_types.h
@@ -222,7 +222,7 @@ typedef struct {
 
 typedef struct {
     sMacAddr mac_adress;
-    uint8_t signal_strength;
+    int32_t signal_strength;
     uint8_t channel;
     uint8_t operating_class;
     uint32_t time_stamp;
